src/851A/test_cdf_851A.py
import unittest
from unittest.mock import patch

from cdf_851A import CodeforcesTask851ASolution


class TestCDF851A(unittest.TestCase):
    def test_851A_acceptance_1(self):
        mock_input = ['10 5 3']
        expected = '3'
        with patch('builtins.input', side_effect=mock_input):
            Solution = CodeforcesTask851ASolution()
            Solution.read_input()
            Solution.process_task()
            actual = Solution.get_result()

        self.assertEqual(expected, actual)

    def test_851A_acceptance_2(self):
        mock_input = ['10 5 7']
        expected = '5'
        with patch('builtins.input', side_effect=mock_input):
            Solution = CodeforcesTask851ASolution()
            Solution.read_input()
            Solution.process_task()
            actual = Solution.get_result()

        self.assertEqual(expected, actual)

    def test_851A_acceptance_3(self):
        mock_input = ['10 5 12']
        expected = '3'
        with patch('builtins.input', side_effect=mock_input):
            Solution = CodeforcesTask851ASolution()
            Solution.read_input()
            Solution.process_task()
            actual = Solution.get_result()

        self.assertEqual(expected, actual)


if __name__ == "__main__":
    unittest.main()
